+++ runtime/Gfx2D.cpp
@@ -9,8 +9,8 @@ extern UISDL *ui;
 extern Console console;
 extern std::atomic_bool ui_started;
 extern std::atomic_bool start_ui;
-extern size_t screen_width;
-extern size_t screen_height;
+extern int screen_width;
+extern int screen_height;
 extern size_t screen_flags;
 
 extern "C" DLLEXTERN void gfx_uicheck() {
